Change database version, andr replace ling to Integer in the add column method o the migration22

+++ app/src/main/java/org/eyeseetea/malariacare/database/migrations/Migration22TreatmentMatrix.java
@@ -29,8 +29,8 @@ public class Migration22TreatmentMatrix extends BaseMigration {
 
     @Override
     public void migrate(SQLiteDatabase database) {
-        addColumn(database, User.class, "organisation", "long");
-        addColumn(database, User.class, "supervisor", "long");
+        addColumn(database, User.class, "organisation", "Integer");
+        addColumn(database, User.class, "supervisor", "Integer");
     }
 
 }
